[Refactor] - Remove duplicate ApiResponse trait
