+++ app/Traits/ApiResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Traits;
 
 trait ApiResponse
 {
